requirements.txt
selenium
webdriver_manager
python-dateutil
openpyxl
natsort
flask
flask-cors
psycopg2-binary